+++ mods/src/compat/internal.h
@@ -7,4 +7,5 @@ MCPI_INTERNAL void _patch_bcm_host_calls();
 MCPI_INTERNAL void _patch_sdl_calls();
 
 // Functionality
+MCPI_INTERNAL extern bool _compat_has_stopped;
 MCPI_INTERNAL void _init_compat_sdl();