+++ mods/src/display-lists/chunks/rebuild/thread.h
@@ -1,0 +1,43 @@
+#pragma once
+
+#include <mods/tesselator/tesselator.h>
+
+#include "CachedLevelSource.h"
+#include "ThreadVector.h"
+
+// Messages Sent To Rebuild Thread
+struct chunk_rebuild_data {
+    int seed = 0;
+    Chunk *chunk = nullptr;
+    CachedLevelSource source;
+    int x0 = 0;
+    int y0 = 0;
+    int z0 = 0;
+    int x1 = 0;
+    int y1 = 0;
+    int z1 = 0;
+};
+MCPI_UNUSED extern ThreadVector _chunks_to_rebuild;
+
+// Messages Received From Rebuild Thread
+static constexpr int num_layers = 3;
+struct rebuilt_chunk_data {
+    Chunk *chunk = nullptr;
+    bool is_empty = true;
+    bool is_layer_empty[num_layers] = {};
+    bool touched_sky = false;
+    const VertexArray<CustomVertexFlat> *vertices[num_layers] = {};
+};
+MCPI_UNUSED extern ThreadVector _rebuilt_chunks;
+
+// Configure Rendering
+MCPI_UNUSED void _configure_tesselator_for_chunk_rebuild(bool enable);
+
+// Biome Data
+MCPI_UNUSED void _init_threaded_biome_source();
+MCPI_UNUSED void _create_biome_source(int seed);
+MCPI_UNUSED void _free_biome_source();
+
+// Start/Stop
+MCPI_UNUSED void _stop_chunk_rebuild_thread();
+MCPI_UNUSED void _start_chunk_rebuild_thread(Level *level);