+++ mods/src/display-lists/internal.h
@@ -1,4 +1,5 @@
 #pragma once
 
-MCPI_INTERNAL void _init_display_lists_chunks();
+MCPI_INTERNAL void _init_display_lists_chunks_rebuild();
+MCPI_INTERNAL void _init_display_lists_chunks_render();
 MCPI_INTERNAL void _init_display_lists_font();